Allow User to Designate Color
This adds an argument to the forceNetwork function that adds a color
property to the data object if it is specified, in the same way that
Links can get Value. If the Color is specified then the output will
color the nodes whatever color is located in the color column. If people
don't use the argument, then the colors default to the color scalar as
normal.